Seguridad creada y aplicada sólo para pruebas (hay que aplicar seguridad a usuarios en el futuro)

+++ src/main/java/duoc/projects/serena/model/UserActiveSession.java
@@ -13,6 +13,5 @@ public class UserActiveSession {
     private Integer sessionId = 1;
 
     @Column(name = "id_usuario_activo", nullable = false)
-    private Integer activeUserId;
+    private Long activeUserId;
 }
-
